% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afficher_barre_verification.R
\name{afficher_barre_verification}
\alias{afficher_barre_verification}
\title{Afficher la barre de vérification}
\usage{
afficher_barre_verification()
}
\value{
Une interface utilisateur Shiny pour la barre de vérification.
}
\description{
Cette fonction génère une interface utilisateur Shiny pour afficher la barre de vérification
avec des boutons interactifs pour sélectionner différents modes d'interaction avec la grille Picross.
}
\examples{
# Afficher la barre de vérification
afficher_barre_verification()

}
