% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afficher_grille_picross.R
\name{afficher_grille_picross}
\alias{afficher_grille_picross}
\title{Afficher la grille Picross avec des boutons et une barre de vérification}
\usage{
afficher_grille_picross(picross)
}
\arguments{
\item{picross}{Une matrice représentant la grille Picross avec des valeurs binaires (0 pour une cellule vide, 1 pour une cellule remplie).}
}
\value{
Une structure HTML représentant la grille Picross avec des boutons interactifs.
}
\description{
Cette fonction génère une représentation visuelle d'une grille de jeu Picross avec des boutons
interactifs pour remplir ou bloquer les cellules, ainsi qu'une barre de vérification pour vérifier
la solution.
}
\examples{
# Générer une grille Picross aléatoire
picross <- generer_puzzle_picross(10)

# Afficher la grille Picross
afficher_grille_picross(picross)

}
