% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afficher_reglement.R
\name{afficher_reglement_text}
\alias{afficher_reglement_text}
\title{Afficher les détails du jeu}
\usage{
afficher_reglement_text()
}
\value{
Les détails du jeu en texte
}
\description{
Cette fonction génère les détails du jeu Picross en texte.
}
