% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afficher_solution.R
\name{afficher_solution}
\alias{afficher_solution}
\title{Afficher la solution du Picross}
\usage{
afficher_solution(solution_matrix, nb_lignes, nb_colonnes)
}
\arguments{
\item{solution_matrix}{Une matrice représentant la solution de la grille Picross avec des valeurs binaires (0 pour une cellule vide, 1 pour une cellule remplie).}

\item{nb_lignes}{Le nombre de lignes de la grille Picross.}

\item{nb_colonnes}{Le nombre de colonnes de la grille Picross.}
}
\value{
Une interface utilisateur Shiny pour afficher la solution du Picross.
}
\description{
Cette fonction modifie la grille existante pour afficher la solution de la grille Picross en colorant les cellules
remplies en vert foncé.
}
\examples{
# Afficher la solution de la grille Picross sur la grille existante
afficher_solution(solution, nb_lignes, nb_colonnes)

}
