% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculer_indices.R
\name{calculer_indices}
\alias{calculer_indices}
\title{Fonction pour calculer les indices d'une ligne ou colonne}
\usage{
calculer_indices(vec)
}
\arguments{
\item{vec}{Le vecteur à partir duquel calculer les indices.}
}
\value{
Un vecteur contenant les longueurs des séquences de 1 dans le vecteur d'entrée.
}
\description{
Cette fonction prend en entrée un vecteur et retourne les longueurs des séquences
de valeurs égales à 1, représentant ainsi les indices des cases remplies dans une ligne
ou une colonne d'une grille de jeu Picross.
}
\examples{
# Exemple d'utilisation avec un vecteur
vec <- c(1, 0, 1, 1, 1, 0, 0, 1, 1)
calculer_indices(vec)
# Résultat : 3 2

}
