% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_picross_ui.R
\name{create_picross_ui}
\alias{create_picross_ui}
\title{Interface utilisateur Shiny pour le Jeu Picross}
\usage{
create_picross_ui(
  theme = shinytheme("superhero"),
  taille_grille = 5,
  min_taille_grille = 5,
  max_taille_grille = 15,
  step_taille_grille = 1
)
}
\arguments{
\item{theme}{Thème visuel à utiliser pour l'application.}

\item{taille_grille}{Taille initiale de la grille Picross.}

\item{min_taille_grille}{Taille minimale autorisée pour la grille Picross.}

\item{max_taille_grille}{Taille maximale autorisée pour la grille Picross.}

\item{step_taille_grille}{Incrément pour la sélection de la taille de la grille Picross.}
}
\value{
Une interface utilisateur Shiny pour le jeu Picross.
}
\description{
Interface utilisateur Shiny pour le Jeu Picross
}
\details{
Cette interface utilisateur Shiny définit la mise en page et les composants graphiques
pour le jeu Picross, y compris la grille, la barre de vérification et les détails du jeu.
}
\examples{
# Créer l'interface utilisateur
ui <- create_picross_ui(theme = "superhero", taille_grille = 5, min_taille_grille = 5, max_taille_grille = 15, step_taille_grille = 5)

}
