% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generer_puzzle_picross.R
\name{generer_puzzle_picross}
\alias{generer_puzzle_picross}
\title{Générer un puzzle Picross aléatoire}
\usage{
generer_puzzle_picross(nb_lignes, nb_colonnes)
}
\arguments{
\item{nb_lignes}{Le nombre de lignes de la grille Picross.}

\item{nb_colonnes}{Le nombre de colonnes de la grille Picross.}
}
\value{
Une matrice binaire représentant le puzzle Picross généré.
}
\description{
Cette fonction crée un puzzle Picross aléatoire en spécifiant le nombre de lignes et de colonnes de la grille.
}
\examples{
# Générer un puzzle Picross de taille 5x5
puzzle <- generer_puzzle_picross(5, 5)

# Générer un puzzle Picross de taille 10x10
puzzle <- generer_puzzle_picross(10, 10)

}
