% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barre_verification_actions.R
\name{masquer_barre_verification}
\alias{masquer_barre_verification}
\title{Masquer la barre de vérification}
\usage{
masquer_barre_verification()
}
\description{
Masquer la barre de vérification
}
