% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrices_10.R
\name{matrices_10}
\alias{matrices_10}
\title{Générer des matrices pour le jeu Picross (Taille 10x10)}
\usage{
matrices_10()
}
\value{
Une liste de matrices pour le jeu Picross avec des formes telles que l'histogramme, un sablier, et le symbole infini.
}
\description{
Cette fonction crée des matrices pour le jeu Picross avec différentes formes en 10x10.
}
\examples{
matrices_jeu_10 <- matrices_10()

}
