% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrices_5.R
\name{matrices_5}
\alias{matrices_5}
\title{Générer des matrices pour le jeu Picross (Taille 5x5)}
\usage{
matrices_5()
}
\value{
Une liste de matrices pour le jeu Picross avec des formes telles qu'une maison, un avion, le chiffre "8", une forme de pique et un cœur.
}
\description{
Cette fonction crée des matrices pour le jeu Picross avec différentes formes en 5x5.
}
\examples{
matrices_jeu_5 <- matrices_5()

}
