% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barre_verification_actions.R
\name{montrer_barre_verification}
\alias{montrer_barre_verification}
\title{Montrer la barre de vérification lorsque la grille est prête}
\usage{
montrer_barre_verification(picross)
}
\arguments{
\item{picross}{La matrice représentant le puzzle Picross}
}
\description{
Montrer la barre de vérification lorsque la grille est prête
}
