% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{server}
\alias{server}
\title{Logique du serveur Shiny pour l'application Picross}
\source{
"generer_puzzle.R"

"afficher_grille.R"

"barre_verification.R"

"afficher_reglement.R"

"barre_verification_actions.R"

"modes.R"

"verifier_grille.R"
}
\usage{
server(input, output)
}
\description{
Ce script définit la logique du serveur Shiny pour l'application Picross. Il charge
différentes fonctions nécessaires à partir de scripts externes et met en œuvre le
comportement de l'application, y compris la génération de la grille, l'affichage,
la vérification et la gestion des modes.
}
