% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verifier_grille.R
\name{verifier_grille}
\alias{verifier_grille}
\title{Vérifier la grille}
\usage{
verifier_grille(picross, input)
}
\arguments{
\item{picross}{La matrice représentant le puzzle Picross}

\item{input}{L'objet input de Shiny}
}
\description{
Vérifier la grille
}
